moved collection related types to types/collection
