+++ src/utils/collection/index.ts
@@ -1,7 +1,5 @@
 import {
     addToCollection,
-    Collection,
-    CollectionType,
     moveToCollection,
     removeFromCollection,
     restoreToCollection,
@@ -15,6 +13,7 @@ import { getData, LS_KEYS } from 'utils/storage/localStorage';
 import { SetDialogMessage } from 'components/MessageDialog';
 import { logError } from 'utils/sentry';
 import constants from 'utils/strings/constants';
+import { Collection, CollectionType } from 'types/collection';
 
 export enum COLLECTION_OPS_TYPE {
     ADD,
